+++ src/main/java/fr/siovision/voyages/infrastructure/dto/gdpr/GdprPasskeyDTO.java
@@ -1,0 +1,9 @@
+package fr.siovision.voyages.infrastructure.dto.gdpr;
+
+import java.time.LocalDateTime;
+
+public record GdprPasskeyDTO(
+        LocalDateTime createdAt,
+        String aaguid
+) {
+}